requirements_dev.txt
coverage==5.3.1
factory-boy==3.2.0
django-stubs==1.8.0
djangorestframework-stubs==1.3.0
mypy==0.790
